abi generation - enum discriminants
