+++ contracts/feature-tests/abi-tester/abi_test_expected.abi.json
@@ -24,6 +24,20 @@
             ]
         },
         {
+            "name": "echo_enum",
+            "inputs": [
+                {
+                    "name": "e",
+                    "type": "AbiEnum"
+                }
+            ],
+            "outputs": [
+                {
+                    "type": "AbiEnum"
+                }
+            ]
+        },
+        {
             "name": "multi_result_3",
             "inputs": [],
             "outputs": [
@@ -142,6 +156,56 @@
         }
     ],
     "types": {
+        "AbiEnum": {
+            "type": "enum",
+            "docs": [
+                "Its only purpose is to test that the ABI generator works fine."
+            ],
+            "variants": [
+                {
+                    "name": "Nothing",
+                    "discriminant": 0
+                },
+                {
+                    "name": "Something",
+                    "discriminant": 1,
+                    "fields": [
+                        {
+                            "name": "0",
+                            "type": "i32"
+                        }
+                    ]
+                },
+                {
+                    "name": "SomethingMore",
+                    "discriminant": 2,
+                    "fields": [
+                        {
+                            "name": "0",
+                            "type": "u8"
+                        },
+                        {
+                            "name": "1",
+                            "type": "OnlyShowsUpAsNested8"
+                        }
+                    ]
+                },
+                {
+                    "name": "SomeStruct",
+                    "discriminant": 3,
+                    "fields": [
+                        {
+                            "name": "a",
+                            "type": "u16"
+                        },
+                        {
+                            "name": "b",
+                            "type": "OnlyShowsUpAsNested9"
+                        }
+                    ]
+                }
+            ]
+        },
         "AbiTestType": {
             "type": "struct",
             "docs": [
@@ -245,6 +309,30 @@
             ]
         },
         "OnlyShowsUpAsNested7": {
+            "type": "struct",
+            "docs": [
+                "Tests that the ABI generator also fetches types that only appear as fields."
+            ],
+            "fields": [
+                {
+                    "name": "something",
+                    "type": "()"
+                }
+            ]
+        },
+        "OnlyShowsUpAsNested8": {
+            "type": "struct",
+            "docs": [
+                "Tests that the ABI generator also fetches types that only appear as fields."
+            ],
+            "fields": [
+                {
+                    "name": "something",
+                    "type": "()"
+                }
+            ]
+        },
+        "OnlyShowsUpAsNested9": {
             "type": "struct",
             "docs": [
                 "Tests that the ABI generator also fetches types that only appear as fields."
