+++ contracts/feature-tests/abi-tester/src/abi_enum.rs
@@ -1,0 +1,14 @@
+use crate::only_nested::*;
+derive_imports!();
+
+/// Its only purpose is to test that the ABI generator works fine.
+#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
+pub enum AbiEnum {
+	Nothing,
+	Something(i32),
+	SomethingMore(u8, OnlyShowsUpAsNested8),
+	SomeStruct{
+		a: u16,
+		b: OnlyShowsUpAsNested9
+	}
+}
