+++ contracts/feature-tests/abi-tester/src/abi_test_type.rs
@@ -1,4 +1,5 @@
 use elrond_wasm::Box;
+use crate::only_nested::*;
 derive_imports!();
 
 /// Its only purpose is to test that the ABI generator works fine.
@@ -15,44 +16,3 @@ pub struct AbiTestType {
 	pub tuple_madness: (OnlyShowsUpAsNested2, Option<Box<AbiTestType>>),
 }
 
-/// Tests that the ABI generator also fetches types that only appear as fields.
-#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
-pub struct OnlyShowsUpAsNested1 {
-	pub something: (),
-}
-
-/// Tests that the ABI generator also fetches types that only appear as fields.
-#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
-pub struct OnlyShowsUpAsNested2 {
-	pub something: (),
-}
-
-/// Tests that the ABI generator also fetches types that only appear as fields.
-#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
-pub struct OnlyShowsUpAsNested3 {
-	pub something: (),
-}
-
-/// Tests that the ABI generator also fetches types that only appear as fields.
-#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
-pub struct OnlyShowsUpAsNested4 {
-	pub something: (),
-}
-
-/// Tests that the ABI generator also fetches types that only appear as fields.
-#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
-pub struct OnlyShowsUpAsNested5 {
-	pub something: (),
-}
-
-/// Tests that the ABI generator also fetches types that only appear as fields.
-#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
-pub struct OnlyShowsUpAsNested6 {
-	pub something: (),
-}
-
-/// Tests that the ABI generator also fetches types that only appear as fields.
-#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
-pub struct OnlyShowsUpAsNested7 {
-	pub something: (),
-}
