+++ contracts/feature-tests/abi-tester/src/only_nested.rs
@@ -1,0 +1,55 @@
+derive_imports!();
+
+/// Tests that the ABI generator also fetches types that only appear as fields.
+#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
+pub struct OnlyShowsUpAsNested1 {
+	pub something: (),
+}
+
+/// Tests that the ABI generator also fetches types that only appear as fields.
+#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
+pub struct OnlyShowsUpAsNested2 {
+	pub something: (),
+}
+
+/// Tests that the ABI generator also fetches types that only appear as fields.
+#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
+pub struct OnlyShowsUpAsNested3 {
+	pub something: (),
+}
+
+/// Tests that the ABI generator also fetches types that only appear as fields.
+#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
+pub struct OnlyShowsUpAsNested4 {
+	pub something: (),
+}
+
+/// Tests that the ABI generator also fetches types that only appear as fields.
+#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
+pub struct OnlyShowsUpAsNested5 {
+	pub something: (),
+}
+
+/// Tests that the ABI generator also fetches types that only appear as fields.
+#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
+pub struct OnlyShowsUpAsNested6 {
+	pub something: (),
+}
+
+/// Tests that the ABI generator also fetches types that only appear as fields.
+#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
+pub struct OnlyShowsUpAsNested7 {
+	pub something: (),
+}
+
+/// Tests that the ABI generator also fetches types that only appear as fields.
+#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
+pub struct OnlyShowsUpAsNested8 {
+	pub something: (),
+}
+
+/// Tests that the ABI generator also fetches types that only appear as fields.
+#[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
+pub struct OnlyShowsUpAsNested9 {
+	pub something: (),
+}
