+++ elrond-wasm-debug/src/abi_json/type_abi_json.rs
@@ -76,6 +76,7 @@ pub struct EnumVariantDescriptionJson {
 	#[serde(skip_serializing_if = "Vec::is_empty")]
 	pub docs: Vec<String>,
 	pub name: String,
+	pub discriminant: usize,
 	#[serde(skip_serializing_if = "Vec::is_empty")]
 	pub fields: Vec<StructFieldDescriptionJson>,
 }
@@ -85,6 +86,7 @@ impl From<&EnumVariantDescription> for EnumVariantDescriptionJson {
 		EnumVariantDescriptionJson {
 			docs: abi.docs.iter().map(|d| d.to_string()).collect(),
 			name: abi.name.to_string(),
+			discriminant: abi.discriminant,
 			fields: abi
 				.fields
 				.iter()
